# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blockchain Freelance Marketplace API",
    "version": "1.0.0",
    "description": "A decentralized freelance marketplace with AI skill matching and blockchain payments",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Contract": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "proposalId": {
            "type": "string",
            "format": "uuid"
          },
          "freelancerId": {
            "type": "string",
            "format": "uuid"
          },
          "employerId": {
            "type": "string",
            "format": "uuid"
          },
          "escrowAddress": {
            "type": "string"
          },
          "totalAmount": {
            "type": "number"
          },
          "baseAmount": {
            "type": "number",
            "description": "Base contract amount before rush fee"
          },
          "rushFee": {
            "type": "number",
            "description": "Rush fee amount (0 if not rush)"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "disputed",
              "cancelled"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Evidence": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "submitterId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "file",
              "link"
            ]
          },
          "content": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DisputeResolution": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "freelancer_favor",
              "employer_favor",
              "split"
            ]
          },
          "reasoning": {
            "type": "string"
          },
          "resolvedBy": {
            "type": "string"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Dispute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contractId": {
            "type": "string"
          },
          "milestoneId": {
            "type": "string"
          },
          "initiatorId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Evidence"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "under_review",
              "resolved"
            ]
          },
          "resolution": {
            "$ref": "#/components/schemas/DisputeResolution"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/contracts": {
      "get": {
        "summary": "List user's contracts",
        "description": "Retrieves all contracts for the authenticated user (as freelancer or employer)",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Number of results per page"
          },
          {
            "in": "query",
            "name": "continuationToken",
            "schema": {
              "type": "string"
            },
            "description": "Token for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Contracts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Contract"
                      }
                    },
                    "hasMore": {
                      "type": "boolean"
                    },
                    "continuationToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/contracts/{id}": {
      "get": {
        "summary": "Get contract details",
        "description": "Retrieves details of a specific contract",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Contract ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Contract retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contract"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Contract not found"
          }
        }
      }
    },
...
    "/api/contracts/{contractId}/disputes": {
      "get": {
        "summary": "List disputes for a contract",
        "description": "Get all disputes associated with a contract",
        "tags": [
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contractId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The contract ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of disputes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dispute"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "User not authorized to view disputes"
          },
          "404": {
            "description": "Contract not found"
          }
        }
      }
    },
...
    "/api/contracts/{id}/fund": {
      "post": {
        "summary": "Fund contract escrow",
        "description": "Employer funds the escrow for a pending contract, activating it",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Contract ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Escrow funded and contract activated"
          },
          "400": {
            "description": "Contract not in pending status or missing wallet addresses"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only the employer can fund the escrow"
          },
          "404": {
            "description": "Contract not found"
          }
        }
      }
    },
    "/api/contracts/{id}/escrow/withdraw": {
      "post": {
        "summary": "Withdraw the employer's pending escrow allocation (real blockchain mode)",
        "description": "The server wallet is the on-chain employer/platform, so this claims its allocation credited by a dispute resolution.",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Contract ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Withdrawal processed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only the employer or an admin can trigger the platform withdrawal"
          },
          "422": {
            "description": "Only available in real blockchain mode"
          }
        }
      }
    },
    "/api/contracts/{id}/cancel": {
      "post": {
        "summary": "Cancel a pending contract",
        "description": "Cancel a contract that is still in pending status",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Contract ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Contract cancelled successfully"
          },
          "400": {
            "description": "Contract cannot be cancelled"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Contract not found"
          }
        }
      }
    }
  },
...
}